requirements.txt
# Core dependencies (pinned for reproducibility)
requests==2.31.0
boto3==1.34.0
botocore==1.34.0
opencv-python==4.9.0.80
numpy==1.26.3

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
responses==0.24.1

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Type stubs for better IDE support
types-requests==2.31.0.20240106
